separating styles from react
